ROSSERIAL (#13)
* Add files via upload

The following files are all related to the auv recruitment tasks. This includes the image processing task, data labelling task and the ROS task

* Add files via upload

The publisher reads the value from the IR Sensor publishes it. The subscriber sub here reads those values and according to it if the values are greater than 5 it makes the LED glows

* Add files via upload